unreal_asset: move fieldpath to kismet module + fix integrator
