+++ unreal_asset/src/kismet/mod.rs
@@ -224,6 +224,21 @@ pub enum EScriptInstrumentationType {
     Stop,
 }
 
+#[derive(Debug, Clone, Default, PartialEq, Eq, Hash)]
+pub struct FieldPath {
+    pub path: Vec<FName>,
+    pub resolved_owner: PackageIndex,
+}
+
+impl FieldPath {
+    pub fn new(path: Vec<FName>, resolved_owner: PackageIndex) -> Self {
+        FieldPath {
+            path,
+            resolved_owner,
+        }
+    }
+}
+
 fn read_kismet_string<Reader: AssetReader>(asset: &mut Reader) -> Result<String, Error> {
     let mut data = Vec::new();
     loop {
